requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.3
django-redis==5.4.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
celery==5.3.4
redis==5.0.1
psutil==5.9.6
mysqlclient==2.2.0
gunicorn==21.2.0
cryptography==41.0.5
boto3==1.28.44
paramiko==3.3.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
pytz==2023.3.post1
pillow==10.1.0
python-magic==0.4.27
dnspython==2.4.2
bcrypt==4.0.1 